requirements.txt
scipy==1.12.0
gensim==4.3.2
redis==5.0.4
aiofiles==24.1.0
slowapi==0.1.9
hiredis==2.3.2
httpx==0.27.0
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
gensim==4.3.2
h11==0.14.0
idna==3.7
joblib==1.4.2
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pypandoc==1.13
pytesseract==0.3.10
python-crfsuite==0.9.10
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scikit-learn==1.5.0
smart-open==7.0.4
sniffio==1.3.1
starlette==0.37.2
threadpoolctl==3.5.0
tqdm==4.66.4
underthesea==6.8.0
underthesea_core==1.0.4
urllib3==2.2.1
uvicorn==0.30.0
wrapt==1.16.0
yarl==1.9.4
python-multipart==0.0.9
